Split Utils class in multiple extension classes
