+++ readxmlfromurl/QAction_5/MergedEntry.cs
@@ -9,8 +9,8 @@ namespace QAction_5
     public class MergedEntry
     {
         public DateTime adSalesTime;
-        public Utils.AdSalesRow adSalesData;
-        public Utils.WhatsonRow whatsonData;
-        public Utils.MediatorRow mediatorData;
+        public AdSalesRow adSalesData;
+        public WhatsonRow whatsonData;
+        public MediatorRow mediatorData;
     }
 }
